Allow reducing the BT connection delay.
By default, a four-second delay is imposed on some devices between
setting up HFP and A2DP.  This makes Bluetooth initialization take
much longer than it needs to for many devices.  This delay is very
annoying; slow Bluetooth connections often mean the user is sitting
around, waiting for it to complete.

Allow disabling this delay if the user's devices don't need it.  I've
assumed that some devices do (it's there for a reason), so the default
is unchanged.

Change-Id: Ibfa0a76b4907cb76ad754c09d15e21b5f8b278cd
Signed-off-by: Glenn Maynard <glenn@zewt.org>